Remove static keys and introduce static_call_cond_int
With the addition of static_call_cond_int to the static call API, we
don't need static keys anymore. A default value is provided, which is
put into %eax before the call. This way, if the call is replaced by a
nop, the default value will be in the place of the returned value.

For now, this does not work when one of the arguments of the call is a
function call (eax will be overwritten), hence the modification to
security_mmap_file